define VP & Pex interfaces and models
