+++ src/Hyperledger.Aries/Features/OpenID4VP/Models/AuthorizationRequest.cs
@@ -1,0 +1,8 @@
+using Hyperledger.Aries.Features.Pex.Models;
+
+namespace Hyperledger.Aries.Features.OpenID4VP.Models;
+
+public class AuthorizationRequest
+{
+    public PresentationDefinition PresentationDefinition { get; set; }
+}
