+++ src/Hyperledger.Aries/Features/OpenID4VP/Models/SelectedCredential.cs
@@ -1,0 +1,11 @@
+using Hyperledger.Aries.Storage;
+
+namespace Hyperledger.Aries.Features.OpenID4VP.Models
+{
+    public class SelectedCredential
+    {
+        public string InputDescriptorId { get; set; }
+        
+        public ICredential Credential { get; set; }
+    }
+}
