+++ src/Hyperledger.Aries/Features/OpenID4VP/Services/IOid4VpClient.cs
@@ -1,11 +1,11 @@
 using System.Threading.Tasks;
-using Hyperledger.Aries.Features.Pex.Models;
+using Hyperledger.Aries.Features.OpenID4VP.Models;
 
 namespace Hyperledger.Aries.Features.OpenID4VP.Services
 {
     public interface IOid4VpClient
     {
-        Task<(CredentialCandidates[], string authorizationRequestRecordId)> ProcessAuthorizationRequest(string authorizationRequestUrl);
+        Task<AuthorizationRequest> ProcessAuthorizationRequest(string authorizationRequestUrl);
   
         Task SendAuthorizationResponse(SelectedCredential[] selectedCredentials, string authorizationRequestRecordId);
     }
