+++ src/Hyperledger.Aries/Features/Pex/Models/InputDescriptor.cs
@@ -1,0 +1,6 @@
+namespace Hyperledger.Aries.Features.Pex.Models;
+
+public class InputDescriptor
+{
+    public string Id { get; set; }
+}
