+++ src/Hyperledger.Aries/Features/Pex/Models/PresentationDefinition.cs
@@ -2,6 +2,6 @@ namespace Hyperledger.Aries.Features.Pex.Models
 {
     public class PresentationDefinition
     {
-        
+        public InputDescriptor[] InputDescriptors { get; set; }
     }
 }
