--- src/Hyperledger.Aries/Features/Pex/Models/SelectedCredential.cs
@@ -1,14 +1,0 @@
-namespace Hyperledger.Aries.Features.Pex.Models
-{
-    public class SelectedCredential
-    {
-        public ICredential CredentialId { get; set; }
-        
-        public string InputDescriptorId { get; set; }
-    }
-
-    public interface ICredential
-    {
-        public string CredentialId { get; set; }
-    }
-}
