+++ src/Hyperledger.Aries/Features/Pex/Services/IPexService.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using Hyperledger.Aries.Features.Pex.Models;
 
@@ -6,22 +5,8 @@ namespace Hyperledger.Aries.Features.Pex.Services
 {
     public interface IPexService
     {
-        //TODO: Method may be extended with JSON Schema (tbd)
         Task<PresentationDefinition> ParsePresentationDefinition(string presentationDefinition);
   
-        Task<PresentationSubmission> CreatePresentationSubmisson(string presentationDefinitionId, CredentialDescriptor[] credentials);
-    }
-
-    public class CredentialDescriptor
-    {
-        public string InputDescriptorId { get; set; }
-        
-        public string CredentialId { get; set; }
-        
-        public string Format { get; set; }
-        
-        public string Path { get; set; }
-        
-        public CredentialDescriptor? PathNested { get; set; }
+        Task<PresentationSubmission> CreatePresentationSubmission(PresentationDefinition presentationDefinition, CredentialDescriptor[] credentials);
     }
 }
