+++ src/Hyperledger.Aries/Features/SdJwt/Services/SdJwtCredentialService/DefaultSdJwtCredentialService.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Hyperledger.Aries.Agents;
 using Hyperledger.Aries.Features.OpenId4Vc.Vci.Models.Metadata.Issuer;
+using Hyperledger.Aries.Features.OpenID4VP.Models;
+using Hyperledger.Aries.Features.Pex.Models;
 using Hyperledger.Aries.Features.SdJwt.Models.Records;
 using Hyperledger.Aries.Storage;
 using SD_JWT.Abstractions;
@@ -73,6 +75,16 @@ namespace Hyperledger.Aries.Features.SdJwt.Services.SdJwtCredentialService
         public virtual async Task<bool> DeleteAsync(IAgentContext context, string recordId)
         {
             return await RecordService.DeleteAsync<SdJwtRecord>(context.Wallet, recordId);
+        }
+
+        public Task<CredentialCandidates[]> GetCredentialCandidates(InputDescriptor[] inputDescriptors)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<string> CreateSdJwtPresentationFormat(InputDescriptor inputDescriptor, string credentialId)
+        {
+            throw new NotImplementedException();
         }
     }
 }
