+++ src/Hyperledger.Aries/Storage/Records/CredentialRecordBase.cs
@@ -1,0 +1,5 @@
+namespace Hyperledger.Aries.Storage;
+
+public interface ICredential
+{
+}
